requirements.txt
# ArgoCD → Sheets (bestaand)
requests==2.32.5
openpyxl==3.1.5

# ISO Audit pipeline
anthropic>=0.40.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
python-docx>=1.1.0
markdown>=3.7
htmldocx>=0.0.6
